+++ src/2d/lib/Curve2D.ts
@@ -51,7 +51,7 @@ export class Curve2D {
     if (!this._deleted) {
       this._deleted = true;
       unregisterFromCleanup(this._wrapped);
-      this._wrapped.delete();
+      if (typeof this._wrapped.delete === 'function') this._wrapped.delete();
     }
   }
